Allow CDR of a number (fixes #7)
Eliza takes the CDR of a number. This is actually legal in BBN-Lisp (see page
5.1 of the Apr69 manual). Commit 841266d broke this by introducing an error
branch in our CDR implementation. Solution: add an explicit NUMBERP test.

This change also broke (CDR NIL), because NIL is not a cons. Change the test
from CONSP to LISTP.

(CONSP is correct for RPLACA and RPLACD, as modifying NIL is not allowed, even
in BBN Lisp (page 3.4))